unify para and seq index build
